refactor: 반복되는 focus 로직 공통화
1. useFocus 훅 추가

2. 비밀번호, cvc 번호, 소유자 이름 TextField 컴포넌트에 적용
